tienda/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==5.1.3
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-icons==0.8.7
django-bootstrap5==24.3
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-jquery==3.1.0
django-seed==0.3.1
django-staticfiles-moment==2.0.0
Faker==33.0.0
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
toposort==1.10
typing_extensions==4.12.2
urllib3==2.2.3
djangorestframework~=3.14.0
requests~=2.31.0
